iot/requirements.txt
paho-mqtt==1.6.1
requests==2.26.0
python-dotenv==1.0.0
AWSIoTPythonSDK==1.5.2